requirements.txt
# FastAPI y servidor
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Google Cloud Platform (GA 2025 stack)
google-cloud-aiplatform==1.82.0
google-cloud-storage==2.16.0
google-cloud-documentai==2.30.0
google-cloud-discoveryengine==0.13.2

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Procesamiento de PDFs e imágenes
PyPDF2==3.0.1
pdf2image==1.17.0
Pillow==11.0.0
pypdfium2==4.26.0

# Procesamiento de texto y embeddings
langchain==0.1.4
langchain-google-vertexai==0.0.6
chromadb==0.4.22
faiss-cpu==1.12.0

# Templates y generación de reportes
jinja2==3.1.3
weasyprint==60.2
markdown==3.5.2

# Utilidades
requests==2.31.0
httpx==0.26.0
aiofiles==23.2.1
python-dateutil==2.8.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Logging y monitoring
structlog==24.1.0

# Rate limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6
